orm+jpa refactor packages and class names
