+++ src/main/java/ru/vitalib/otus/homework/books/dao/AuthorDaoJPA.java
@@ -1,4 +1,4 @@
-package ru.vitalib.otus.homework.books.repositories;
+package ru.vitalib.otus.homework.books.dao;
 
 import java.util.List;
 import javax.persistence.EntityManager;
@@ -10,11 +10,11 @@ import org.springframework.stereotype.Repository;
 import ru.vitalib.otus.homework.books.domain.Author;
 
 @Repository
-public class AuthorRepositoryJPA implements AuthorRepository {
+public class AuthorDaoJPA implements AuthorDao {
   @PersistenceContext
   private final EntityManager em;
 
-  public AuthorRepositoryJPA(EntityManager em) {
+  public AuthorDaoJPA(EntityManager em) {
     this.em = em;
   }
