+++ src/main/java/ru/vitalib/otus/homework/books/dao/GenreDao.java
@@ -1,9 +1,9 @@
-package ru.vitalib.otus.homework.books.repositories;
+package ru.vitalib.otus.homework.books.dao;
 
 import java.util.List;
 import ru.vitalib.otus.homework.books.domain.Genre;
 
-public interface GenreRepository {
+public interface GenreDao {
   Genre save(Genre author);
 
   Genre findById(long id);
