+++ src/main/java/ru/vitalib/otus/homework/books/dao/GenreDaoJPA.java
@@ -1,4 +1,4 @@
-package ru.vitalib.otus.homework.books.repositories;
+package ru.vitalib.otus.homework.books.dao;
 
 import java.util.List;
 import javax.persistence.EntityManager;
@@ -8,12 +8,12 @@ import org.springframework.stereotype.Repository;
 import ru.vitalib.otus.homework.books.domain.Genre;
 
 @Repository
-public class GenreRepositoryJPA implements GenreRepository {
+public class GenreDaoJPA implements GenreDao {
 
   @PersistenceContext
   private final EntityManager em;
 
-  public GenreRepositoryJPA(EntityManager em) {
+  public GenreDaoJPA(EntityManager em) {
     this.em = em;
   }
