+++ src/main/java/ru/vitalib/otus/homework/books/service/SimpleAuthorService.java
@@ -5,16 +5,16 @@ import lombok.AllArgsConstructor;
 import org.springframework.stereotype.Service;
 import ru.vitalib.otus.homework.books.domain.Author;
 import ru.vitalib.otus.homework.books.exception.AuthorNotFoundException;
-import ru.vitalib.otus.homework.books.repositories.AuthorRepository;
+import ru.vitalib.otus.homework.books.dao.AuthorDao;
 
 @Service
 @AllArgsConstructor
 public class SimpleAuthorService implements AuthorService {
 
-  private final AuthorRepository authorRepository;
+  private final AuthorDao authorDao;
 
   @Override
   public Author getAuthorByName(String authorName) {
-    return Optional.ofNullable(authorRepository.findByName(authorName)).orElseThrow(AuthorNotFoundException::new);
+    return Optional.ofNullable(authorDao.findByName(authorName)).orElseThrow(AuthorNotFoundException::new);
   }
 }
