+++ src/main/java/ru/vitalib/otus/homework/books/service/SimpleGenreService.java
@@ -6,15 +6,15 @@ import ru.vitalib.otus.homework.books.domain.Genre;
 import ru.vitalib.otus.homework.books.exception.GenreNotFoundException;
 
 import java.util.Optional;
-import ru.vitalib.otus.homework.books.repositories.GenreRepository;
+import ru.vitalib.otus.homework.books.dao.GenreDao;
 
 @Service
 @AllArgsConstructor
 public class SimpleGenreService implements GenreService {
-  public final GenreRepository genreRepository;
+  public final GenreDao genreDao;
 
   @Override
   public Genre getGenreByName(String genreName) {
-    return Optional.ofNullable(genreRepository.findByName(genreName)).orElseThrow(GenreNotFoundException::new);
+    return Optional.ofNullable(genreDao.findByName(genreName)).orElseThrow(GenreNotFoundException::new);
   }
 }
