+++ src/test/java/ru/vitalib/otus/homework/books/dao/GenreDaoJPATest.java
@@ -1,4 +1,4 @@
-package ru.vitalib.otus.homework.books.repositories;
+package ru.vitalib.otus.homework.books.dao;
 
 import java.util.List;
 import static org.assertj.core.api.Assertions.assertThat;
@@ -8,12 +8,13 @@ import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.test.autoconfigure.orm.jpa.DataJpaTest;
 import org.springframework.boot.test.autoconfigure.orm.jpa.TestEntityManager;
 import org.springframework.context.annotation.Import;
+import ru.vitalib.otus.homework.books.dao.GenreDaoJPA;
 import ru.vitalib.otus.homework.books.domain.Genre;
 
 @DisplayName("GenreDao test")
 @DataJpaTest
-@Import(GenreRepositoryJPA.class)
-class GenreRepositoryJPATest {
+@Import(GenreDaoJPA.class)
+class GenreDaoJPATest {
 
   @Autowired
   private TestEntityManager em;
@@ -22,7 +23,7 @@ class GenreRepositoryJPATest {
   public static final Genre EXISTING_GENRE = new Genre(1, "Детектив");
 
   @Autowired
-  private GenreRepositoryJPA genreRepositoryJPA;
+  private GenreDaoJPA genreRepositoryJPA;
 
   @Test
   @DisplayName("Find genre by id")
